Initial support for rendering.
- implementation and tests for camera
- preferences for axis colors
- transformation class
- render context attributes
- add spec for position / color vertices to vertex array
- some minor fixes

+++ test/src/MatTest.cpp
@@ -680,7 +680,7 @@ TEST(MatTest, write) {
                      5.0,  6.0,  7.0,  8.0,
                      9.0, 10.0, 11.0, 12.0,
                     13.0, 14.0, 15.0, 16.0);
-    float buffer[16];
+    double buffer[16];
     m.write(buffer);
     for (size_t c = 0; c < 4; ++c) {
         for (size_t r = 0; r < 3; ++r) {
